Viewport is floating point to match D3D11 and OpenGL 4.1; added GraphicsDevice::GetViewport(); added automatic mipmap generation; added NXNA_DEPTHSTENCIL_DEPTHREAD macro; added SpriteBatch sample

+++ branches/Nxna2/samples/SpinningTriangle/main.cpp
@@ -58,7 +58,7 @@ int main(int argc, char** argv)
 	device->SetMessageCallback(msgCallback);
 
 	// set the viewport
-	Nxna::Graphics::Viewport vp(0, 0, screenWidth, screenHeight);
+	Nxna::Graphics::Viewport vp(0, 0, (float)screenWidth, (float)screenHeight);
 	device->SetViewport(vp);
 
 	// create the matrices
